updated item model and added to seed.js
