+++ server/db/models/Item.js
@@ -7,15 +7,27 @@ const Item = db.define('item', {
     unique: true,
     allowNull: false
   },
-  description: Sequelize.TEXT,
+
+  imgUrl: {
+    type: Sequelize.STRING,
+    defaultValue: 'https://pbs.twimg.com/profile_images/3100168478/20aea1ef732e8c5e68634dcb8236b4ef.jpeg'
+  },
+
+  description: {
+    type: Sequelize.TEXT,
+    defaultValue: 'This item has no description!'
+  },
 
   quantity: {
     type: Sequelize.INTEGER,
-    min: 0
+    min: 0,
+    defaultValue: 0
   },
+
   price: {
       type: Sequelize.FLOAT,
-      min: 0
+      min: 0,
+      defaultValue: 0
   }
 })
